+++ pose_estimation.py
@@ -2,6 +2,7 @@ import mediapipe as mp
 import cv2
 import numpy as np
 import time 
+from get_angles import get_angles
 
 
 '''
@@ -26,18 +27,6 @@ rightShoulderStage = None
 
 finishedLeftPunch = False
 finishedRightPunch = False
-
-#this takes three points on the pose net, like elbow, shoulder, and wrist, and calculates the angle that there is there
-def calculate_angle(a, b, c):
-    a = np.array(a)
-    b = np.array(b)
-    c = np.array(c)
-
-    radians = np.arctan2(c[1] - b[1], c[0] - b[0]) - np.arctan2(a[1] - b[1], a[0] - b[0])
-    angle = np.abs(radians * 180.0 / np.pi)
-    if angle > 180.0:
-        angle = 360 - angle
-    return angle
 
 
 
@@ -95,53 +84,7 @@ with mp_holistic.Holistic(min_detection_confidence=0.3, min_tracking_confidence=
 
 
         try:
-            landmarks = results.pose_landmarks.landmark
-            #.landmark gives you hashmap stuff that has references to all of your xyz tuples 
-
-            # get coords for relevant points
-
-            #get the x and y of the given body point on the 2d camera plane
-
-            '''
-             mp_holistic.PoseLandmark.LEFT_SHOULDER.value,
-    mp_holistic.PoseLandmark.LEFT_ELBOW.value,
-    mp_holistic.PoseLandmark.LEFT_WRIST.value,
-    mp_holistic.PoseLandmark.LEFT_HIP.value,
-
-    # right side
-    mp_holistic.PoseLandmark.RIGHT_SHOULDER.value,
-    mp_holistic.PoseLandmark.RIGHT_ELBOW.value,
-    mp_holistic.PoseLandmark.RIGHT_WRIST.value,
-    mp_holistic.PoseLandmark.RIGHT_HIP.value,
-            '''
-
-            # landmarks is similar to a list of hash-maps: with each coordinate in the list: {12:{'x': 3.141, 'y': 2.712, 'z': 12.31},... }
-
-            # these are all coordinates (x, y)
-
-            left_shoulder = [landmarks[relevant_landmarks_numerical[0]].x, landmarks[relevant_landmarks_numerical[0]].y]
-            left_elbow = [landmarks[relevant_landmarks_numerical[1]].x, landmarks[relevant_landmarks_numerical[1]].y]
-            left_wrist = [landmarks[relevant_landmarks_numerical[2]].x, landmarks[relevant_landmarks_numerical[2]].y]
-            left_hip = [landmarks[relevant_landmarks_numerical[3]].x, landmarks[relevant_landmarks_numerical[3]].y]
-
-            right_shoulder = [landmarks[relevant_landmarks_numerical[4]].x, landmarks[relevant_landmarks_numerical[4]].y]
-            right_elbow = [landmarks[relevant_landmarks_numerical[5]].x, landmarks[relevant_landmarks_numerical[5]].y]
-            right_wrist = [landmarks[relevant_landmarks_numerical[6]].x, landmarks[relevant_landmarks_numerical[6]].y]
-            right_hip = [landmarks[relevant_landmarks_numerical[7]].x, landmarks[relevant_landmarks_numerical[7]].y]
-
-            # calculate angle between relevant points
-            leftBicepAngle = calculate_angle(left_shoulder, left_elbow, left_wrist)
-            rightBicepAngle = calculate_angle(right_shoulder, right_elbow, right_wrist)
-
-            leftShoulderAngle = calculate_angle(left_hip, left_shoulder, left_elbow)
-            rightShoulderAngle = calculate_angle(right_hip, right_shoulder, right_elbow)
-
-            
-            #print(angle)
-            #time.sleep(0.5)
-
-
-            # tracks if a bicep curl has been completed and how many bicep curls have been completed
+            leftBicepAngle, leftShoulderAngle, rightBicepAngle, rightShoulderAngle = get_angles(image, holistic, relevant_landmarks_numerical)
 
 
             # Biceps
@@ -160,83 +103,32 @@ with mp_holistic.Holistic(min_detection_confidence=0.3, min_tracking_confidence=
 
             #setting shoulderstage
             punched = leftShoulderStage and leftBicepStage
+
+            print(punched)
             if punched:
-                continue
+                print('counter += 1')
+                while punched:
+                    leftBicepAngle, leftShoulderAngle, rightBicepAngle, rightShoulderAngle = get_angles(image, holistic, relevant_landmarks_numerical)
+                    # Biceps
+                    if leftBicepAngle < 25:
+                        leftBicepStage = False
+                    elif leftBicepAngle > 150 and not leftBicepStage:
+                        leftBicepStage = True
 
-            '''
-            if rightBicepAngle < 25 or finishedRightPunch:
-                rightBicepStage = False
-                finishedRightPunch = False
-            elif rightBicepAngle > 150 and not rightBicepStage and not finishedRightPunch:
-                rightBicepStage = True
+                    #in this case, we set our bicep stage
 
+                    # Shoulders
+                    if leftShoulderAngle < 15:
+                        leftShoulderStage = False
+                    elif leftShoulderAngle > 30 and not leftShoulderStage:
+                        leftShoulderStage = True
 
+                    #setting shoulderstage
+                    punched = leftShoulderStage and leftBicepStage
 
-            list = [
-            left_elbow_angle at time t = 0 is the 0th element,
-            left_elbow_angle at time t = 1 is the 1th element.
-            ]
-
-            math = (list[1] - list[2]) / (time difference between frames) = rate of change of the elbow angle
-
-            if math > some number:
-            
-            if it is + then the elbow angle is increasing
-
-            if it is 0 then the elbow angle is not changing
-
-            if it is - then im retracting my arm
-
-
-
-            '''
-
-           
-
-            '''
-            if rightShoulderAngle < 15 or finishedRightPunch:
-                rightShoulderStage = False
-                finishedRightPunch = False
-            elif rightShoulderAngle > 30 and not rightShoulderStage:
-                rightShoulderStage = True
-            '''
-
-            # random logs
-            #print(f'Right Shoulder Sage: {rightShoulderAngle}, {rightShoulderStage}')
-            #print(f'Right Bicep Stage: {rightBicepAngle}, {rightBicepStage}')
-            #print(f'Finished Right Punch: {finishedRightPunch}')
-
-            #print('Right Bicep Angle: ', {rightBicepAngle})
-
-            #time.sleep(0.5)
-
-
-            if leftBicepStage and leftShoulderStage:
-                leftBicepStage = None
-                leftShoulderStage = None
-                counter += 1
-                print('left punch: ', counter)
-
-
-            '''
-            if rightBicepStage and rightShoulderStage and not finishedRightPunch:
-                rightBicepStage = None
-                rightShoulderStage = None
-                finishedRightPunch = True
-                counter += 1
-                print('right punch: ', counter)
-            '''
-            
 
         except:
             pass
-
-
-        # render / draw landmarks
-
-        # face stuff
-        # mp_drawing.draw_landmarks(image, results.face_landmarks, mp_holistic.FACEMESH_TESSELATION)
-        # mp_drawing.draw_landmarks(image, results.face_landmarks, mp_holistic.FACEMESH_CONTOURS)
 
         # hand stuff
